+++ lib/malison.dart
@@ -4,5 +4,5 @@ export 'src/canvas_terminal.dart';
 export 'src/glyph.dart';
 export 'src/key_bindings.dart';
 export 'src/retro_terminal.dart';
-export 'src/screen.dart';
 export 'src/terminal.dart';
+export 'src/user_interface.dart';